+++ spec/custom-converter.ts
@@ -1,43 +1,44 @@
 import {expect} from 'chai';
 import {deserialize, serialize, JsonProperty, ICustomConverter} from '../index';
+import dateConverter from './common/dateconverter';
 
 
-const dateConverter: ICustomConverter = {
-    fromJson(data: any): any {
-        return new Date(data);
-    },
+// const dateConverter: ICustomConverter = {
+//     fromJson(data: any): any {
+//         return new Date(data);
+//     },
+//
+//     toJson(data: any): any {
+//         return 'some-date';
+//     }
+// };
 
-    toJson(data: any): any {
-        return 'some-date';
-    }
-};
-
-class Student {
-
-    @JsonProperty('Name')
-    name: string = undefined;
-
-    @JsonProperty({customConverter: dateConverter})
-    dateOfBirth: Date = undefined;
-
-    constructor() {}
-}
+// class Student {
+//
+//     @JsonProperty('Name')
+//     name: string = undefined;
+//
+//     @JsonProperty({customConverter: dateConverter})
+//     dateOfBirth: Date = undefined;
+//
+//     constructor() {}
+// }
 
 
 //TODO put in main test file where deserialization is tested
-describe('custom-converter', function () {
-
-    it('should use the custom converter if available for deserialization', function () {
-        const json = {
-            "Name": "Mark",
-            dateOfBirth: "1995-11-10"
-        };
-        const student = deserialize(Student, json);
-        expect(student.name).to.be.equals('Mark');
-        expect(student.dateOfBirth).to.be.instanceof(Date);
-    });
-
-});
+// describe('custom-converter', function () {
+//
+//     it('should use the custom converter if available for deserialization', function () {
+//         const json = {
+//             "Name": "Mark",
+//             dateOfBirth: "1995-11-10"
+//         };
+//         const student = deserialize(Student, json);
+//         expect(student.name).to.be.equals('Mark');
+//         expect(student.dateOfBirth).to.be.instanceof(Date);
+//     });
+//
+// });
 
 
 describe('serialize', function () {
